chore: Rename remote module packages to be consistent
Previously, the module name was :data:remote but the package names did not reflect this
