+++ data/remote/src/main/kotlin/com/routesearch/data/remote/area/search/AreaSearchDataSource.kt
@@ -1,4 +1,4 @@
-package com.routesearch.network.area.search
+package com.routesearch.data.remote.area.search
 
 interface AreaSearchDataSource {
