+++ data/remote/src/main/kotlin/com/routesearch/data/remote/area/search/AreaSearchResult.kt
@@ -1,4 +1,4 @@
-package com.routesearch.network.area.search
+package com.routesearch.data.remote.area.search
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
