+++ data/src/main/kotlin/com/routesearch/data/DataModule.kt
@@ -1,7 +1,7 @@
 package com.routesearch.data
 
+import com.routesearch.data.remote.remoteDataModule
 import com.routesearch.local.localDataModule
-import com.routesearch.network.remoteDataModule
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
