+++ data/src/main/kotlin/com/routesearch/data/area/search/AreaSearchResultMapper.kt
@@ -1,6 +1,6 @@
 package com.routesearch.data.area.search
 
-import com.routesearch.network.area.search.AreaSearchResult as RemoteAreaSearchResult
+import com.routesearch.data.remote.area.search.AreaSearchResult as RemoteAreaSearchResult
 
 internal fun List<RemoteAreaSearchResult>.toAreaSearchResults() = map {
   it.toAreaSearchResult()
